requirements.txt
torch 
numpy 
regex 

datasets 

pyyaml 

tqdm 
psutil 

matplotlib
